[#1238] Split up hono.app by introducing health check config properties
