+++ tests/src/test/resources/auth/application.yml
@@ -1,8 +1,9 @@
 hono:
   app:
     maxInstances: 1
-    healthCheckBindAddress: 0.0.0.0
-    healthCheckPort: ${vertx.health.port}
+  healthCheck:
+    bindAddress: 0.0.0.0
+    port: ${vertx.health.port}
   auth:
     amqp:
       insecurePortEnabled: true
