+++ tests/src/test/resources/deviceregistry/application.yml
@@ -1,9 +1,10 @@
 hono:
   app:
     maxInstances: 1
-    healthCheckBindAddress: 0.0.0.0
-    healthCheckPort: ${vertx.health.port}
     startupTimeout: 90
+  healthCheck:
+    bindAddress: 0.0.0.0
+    port: ${vertx.health.port}
   auth:
     host: ${hono.auth.host}
     port: 5672
